pyproject.toml
[project]
name = "pax-detector"
version = "0.1.4 "
description = "A vision pipeline for object and car make classification with CLI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Gabriel Ayres", email = "gabn2012@gmail.com" } 
]
keywords = ["computer-vision", "object-detection", "cli", "yolo", "vit"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "datasets>=4.0.0",
    "fastapi[standard]>=0.116.1",
    "gradio>=5.44.1",
    "huggingface-hub>=0.34.4",
    "opencv-python>=4.11.0.86",
    "pathlib>=1.0.1",
    "pillow>=11.3.0",
    "pytorch-lightning>=2.5.3",
    "scikit-learn>=1.7.1",
    "tensorboardx>=2.6.4",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchmetrics>=1.8.1",
    "transformers>=4.55.4",
    "ultralytics>=8.3.188",
    "uvicorn>=0.35.0",
]

[project.urls]
"Homepage" = "https://github.com/gabrielfnayres/pax-case"
"Repository" = "https://github.com/gabrielfnayres/pax-case" 

[project.scripts]
detector = "cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
